Create a nucypher.device module and place TrustedDevice in base.py
